refactor: 调整 auth 文件结构
