+++ src/auth/dtos/auth-response.dto.ts
@@ -1,5 +1,5 @@
 import { ApiResponse } from '../../common/dto/api-response.dto';
-import { AuthData, RefreshTokenData } from '../types/auth.types';
+import { AuthData, RefreshTokenData } from '../type/auth.types';
 
 export type AuthResponseDto = ApiResponse<AuthData>;
 export type RefreshTokenResponseDto = ApiResponse<RefreshTokenData>;
