--- src/auth/strategy/local.strategy.ts
@@ -1,21 +1,0 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { PassportStrategy } from '@nestjs/passport';
-import { Strategy } from 'passport-local';
-import { AuthService } from '../auth.service';
-import { users } from '@prisma/client';
-
-@Injectable()
-export class LocalStrategy extends PassportStrategy(Strategy) {
-    constructor(private authService: AuthService) {
-        super({
-            usernameField: 'email',
-        });
-    }
-    async validate(email: string, password: string): Promise<users> {
-        const user = await this.authService.validateUser(email, password);
-        if (!user) {
-            throw new UnauthorizedException();
-        }
-        return user;
-    }
-}
